eface/e-face/backend/requirements.txt
fastapi
uvicorn[standard]==0.22.0
python-multipart==0.0.6
jinja2==3.1.2
PyJWT==2.8.0
requests==2.31.0
httpx==0.27.0
websockets==11.0.3
